embassy_embedded_hal::adapter
Struct BlockingAsync

pub struct BlockingAsync<T> { /* private fields */ }

Wrapper that implements async traits using blocking implementations.

This allows driver writers to depend on the async traits while still supporting embedded-hal peripheral implementations.

BlockingAsync will implement any async trait that maps to embedded-hal traits implemented for the wrapped driver.

Driver users are then free to choose which implementation that is available to them.

Implementations

impl<T> BlockingAsync<T>

pub fn new(wrapped: T) -> Self

Create a new instance of a wrapper for a given peripheral.